requirements.txt
asgiref==3.10.0
dj-database-url==2.2.0
Django==5.2.7
django-widget-tweaks==1.5.0
gunicorn==23.0.0
packaging==25.0
pillow==12.0.0
psycopg==3.2.12
psycopg-binary==3.2.12
python-dotenv==1.2.1
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
whitenoise==6.11.0
djangorestframework
djangorestframework-simplejwt
